factor out shared textarea event listener into generic
